+++ plugins/html-report-wakamiti-plugin/src/main/java/es/iti/wakamiti/report/html/HtmlReportGenerator.java
@@ -167,7 +167,4 @@ public class HtmlReportGenerator implements Reporter {
         return classLoader().getResourceAsStream(resource);
     }
 
-    private String version() {
-        return getClass().getPackage().getImplementationVersion();
-    }
 }